requirements.txt
anthropic>=0.79.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
lxml>=4.9.0
mcp>=1.0.0
